Improve: Extract generation_manager from tokenizer_manager (#3115)
